// frontend/src/types/api/listBlocklistApiSpamBlocklistGetParams.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */
import type { BlocklistEntryType } from './blocklistEntryType';

export type ListBlocklistApiSpamBlocklistGetParams = {
/**
 * Search by value
 */
search?: string | null;
/**
 * Filter by type: email, domain, pattern
 */
entry_type?: BlocklistEntryType | null;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
per_page?: number;
};