// frontend/src/types/api/blocklistEntryType.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

/**
 * Type of blocklist entry.
 */
export type BlocklistEntryType = typeof BlocklistEntryType[keyof typeof BlocklistEntryType];


export const BlocklistEntryType = {
  email: 'email',
  domain: 'domain',
  pattern: 'pattern',
} as const;